Add harpoon and gen
Added harpoon an gen plugins, along with relevant keybinds. Note that gen does not currently work within the editor,
however, it can be instantly run in a popup terminal, to interact with as normal. Also note that it currently uses codellama,
but this may be changed to mistral due to mistral's power.